requirements.txt
alembic==1.6.5
apispec==5.0.0
appdirs==1.4.4
attrs==21.2.0
backports.entry-points-selectable==1.1.0
black==21.7b0
cfgv==3.3.0
click==8.0.1
distlib==0.3.2
factory-boy==3.2.0
Faker==8.11.0
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-JWT-Extended==4.2.3
Flask-Migrate==3.1.0
flask-smorest==0.32.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
identify==2.2.13
iniconfig==1.1.1
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.4
MarkupSafe==2.0.1
marshmallow==3.13.0
mccabe==0.6.1
mypy-extensions==0.4.3
nodeenv==1.6.0
packaging==21.0
passlib==1.7.4
pathspec==0.9.0
platformdirs==2.2.0
pluggy==0.13.1
pre-commit==2.14.0
psycopg2==2.9.1
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==2.1.0
pyparsing==2.4.7
pytest==6.2.4
pytest-flask==1.2.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-editor==1.0.4
PyYAML==5.4.1
regex==2021.8.3
six==1.16.0
SQLAlchemy==1.4.22
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
unify==0.5
untokenize==0.1.1
virtualenv==20.7.1
webargs==8.0.0
Werkzeug==2.0.1
